Ensure that this gem plays nice with the Railties for ActionController
Including this gem completely breaks view testing, as it causes the
`ActionView::TestCase` class to be loaded before the railtie hooks are
actually defined. This is a problem, because the Action Controller
Railtie extends several modules which override the `inherited` method,
to then redefine the `_routes` method on the new subclass to include a
new anonymous module, which appears to also be the same as the module
causing the `_routes` method to be redefined.

All of this leads to routing helpers not being defined on the view
class in tests. This appears to affect both controller and view tests.
After this change, the classes will be loaded as normal.